Fix runtime issues

+++ src/light.cpp
@@ -17,7 +17,7 @@ Light::Light(Entity *entity, Graphics &gfx, int num, float scale)
 
 	memset(quad_tex, 0, sizeof(unsigned int) * 6);
 	memset(depth_tex, 0, sizeof(unsigned int) * 6);
-	generate_cubemaps(1024 * scale, 1024 * scale);
+	generate_cubemaps((int)(1024 * scale), (int)(1024 * scale));
 }
